+++ examples/test/directives/test_directives_bad_double_else.pas
@@ -1,0 +1,7 @@
+program TestDirectivesBadDoubleElse;
+{$IFDEF FOO}
+{$ELSE}
+{$ELSE}
+{$ENDIF}
+begin
+end.