+++ examples/test/directives/test_directives_bad_noendif.pas
@@ -1,0 +1,5 @@
+program TestDirectivesBadNoEndif;
+{$IFDEF FOO}
+begin
+    writeln('never gets here');
+end.