+++ examples/test/directives/test_directives_bad_stray_else.pas
@@ -1,0 +1,4 @@
+program TestDirectivesBadStrayElse;
+{$ELSE}
+begin
+end.