+++ examples/test/directives/test_directives_bad_stray_endif.pas
@@ -1,0 +1,4 @@
+program TestDirectivesBadStrayEndif;
+{$ENDIF}
+begin
+end.