+++ examples/test/directives/test_directives_basic.pas
@@ -1,0 +1,7 @@
+program TestDirectivesBasic;
+{$DEFINE DEBUG}
+begin
+{$IFDEF DEBUG}
+    writeln('debug on');
+{$ENDIF}
+end.