+++ examples/test/directives/test_directives_dead_code_skipped.pas
@@ -1,0 +1,7 @@
+program TestDirectivesDeadCodeSkipped;
+begin
+{$IFDEF NEVER_DEFINED}
+    ThisIdentifierDoesNotExistAtAll := +++ ;;; garbage nonsense here;
+{$ENDIF}
+    writeln('compiled fine despite garbage in dead branch');
+end.