+++ examples/test/directives/test_directives_ifndef.pas
@@ -1,0 +1,6 @@
+program TestDirectivesIfndef;
+begin
+{$IFNDEF NOPE}
+    writeln('NOPE is not defined');
+{$ENDIF}
+end.