+++ examples/test/directives/test_directives_undef.pas
@@ -1,0 +1,10 @@
+program TestDirectivesUndef;
+{$DEFINE X}
+{$UNDEF X}
+begin
+{$IFDEF X}
+    writeln('should not print');
+{$ELSE}
+    writeln('X undefined now');
+{$ENDIF}
+end.