+++ examples/test/directives/test_directives_units_persist.pas
@@ -1,0 +1,6 @@
+program TestDirectivesUnitsPersist;
+{$DEFINE FROM_MAIN}
+uses DirUnit;
+begin
+    Report;
+end.