+++ examples/test/directives/test_directives_unrecognized.pas
@@ -1,0 +1,7 @@
+program TestDirectivesUnrecognized;
+{$R+}
+begin
+    writeln('R+ accepted and ignored');
+{$Q-}
+    writeln('Q- accepted and ignored');
+end.